Ajout d'utilisateur, de 5 microbrasseries, de 25 bières et 16 sortes de bières parents dans la BD

+++ scripts/creation_tables.sql
@@ -54,7 +54,7 @@ CREATE TABLE Biere (
     prix FLOAT(2),
     nom VARCHAR(100),
     pourcentage_alcool FLOAT(1),
-    ibu FLOAT(1),
+    ibu INTEGER,
     annee_production INTEGER,
     mention VARCHAR(200),
     id_microbrasserie INTEGER NOT NULL,
